% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR-I-Score.R
\name{combine2Forests}
\alias{combine2Forests}
\title{Combine two projection forests}
\usage{
combine2Forests(mod1, mod2)
}
\arguments{
\item{mod1}{A fitted forest object.}

\item{mod2}{A fitted forest object.}
}
\value{
A forest object containing trees from both input forests.
}
\description{
Combine two projection forests
}
\keyword{internal}
