% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR-I-Score.R
\name{combineForests}
\alias{combineForests}
\title{Combine a list of forests}
\usage{
combineForests(list.rf)
}
\arguments{
\item{list.rf}{A list of fitted forest objects.}
}
\value{
A single forest object obtained by combining all forests in \code{list.rf}.
}
\description{
Combine a list of forests
}
\keyword{internal}
