% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-Score.R
\name{compare_Iscores}
\alias{compare_Iscores}
\title{Calculates IScores for multiple imputation functions}
\usage{
compare_Iscores(X, methods_list, score = c("energy_IScore", "DR_IScore"), ...)
}
\arguments{
\item{X}{data containing missing values denoted with NA's.}

\item{methods_list}{a named list of imputing functions.}

\item{score}{a vector of names of scores to calculate. It can be
\code{"energy_IScore"} and \code{"DR_IScore"}.}

\item{...}{other arguments to be passed to  \link[Iscores]{energy_IScore} or
\link[Iscores]{DR_IScore}}
}
\value{
a vector of IScores for provided methods
}
\description{
Calculates IScores for multiple imputation functions
}
\examples{
set.seed(111)
X <- random_mcar_data(100, 3, 0.2)
methods_list <- list(exp = exp_imputation,
                       norm = norm_imputation)
compare_Iscores(X, methods_list = methods_list, m = 2,
                n_proj = 10, n_trees_per_proj = 2 )

}
