% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR-I-Score.R
\name{compute_drScore}
\alias{compute_drScore}
\title{Compute the density ratio score}
\usage{
compute_drScore(object, Z = Z, n_trees_per_proj, n_proj)
}
\arguments{
\item{object}{a crf object.}

\item{Z}{a matrix of candidate points.}

\item{n_trees_per_proj}{an integer, the number of trees per projection.}

\item{n_proj}{an integer specifying the number of projections.}
}
\value{
a numeric value, the DR I-Score.
}
\description{
Compute the density ratio score
}
\keyword{internal}
