% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR-I-Score.R
\name{densityRatioScore}
\alias{densityRatioScore}
\title{Computation of the density ratio score}
\usage{
densityRatioScore(
  X,
  X_imp,
  pattern = NULL,
  n_proj = 10,
  n_trees_per_proj = 1,
  projection_function = NULL,
  min_node_size = 1,
  normal_proj = TRUE
)
}
\arguments{
\item{X}{A numeric matrix of observed data that may contain missing values
denoted by \code{NA}.}

\item{X_imp}{A numeric matrix of imputed values with the same dimensions as
\code{X}.}

\item{pattern}{A vector or pattern indicating the missingness structure.}

\item{n_proj}{An integer specifying the number of random projections.}

\item{n_trees_per_proj}{An integer specifying the number of trees grown per
projection.}

\item{projection_function}{A function that generates user-defined projections.}

\item{min_node_size}{An integer specifying the minimum number of observations
in a terminal node (leaf) of each tree.}

\item{normal_proj}{Logical. If \code{TRUE}, sampling is performed from both
missing (NA) and observed values. If \code{FALSE}, sampling is performed only
from missing (NA) values.}
}
\value{
An object representing a fitted random forest model based on random
projections.
}
\description{
Computes the density ratio score using a random forest model
based on random projections.
}
\details{
The method builds multiple random forests on projected versions of
the data to estimate the density ratio between observed and imputed
distributions.
}
\keyword{internal}
