% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR-I-Score.R
\name{DR_IScore}
\alias{DR_IScore}
\title{Compute the imputation KL-based scoring rules}
\usage{
DR_IScore(
  X,
  imputation_func = NULL,
  X_imp = NULL,
  m = 5,
  n_proj = 100,
  n_trees_per_proj = 5,
  min_node_size = 10,
  n_cores = 1,
  projection_function = NULL,
  ...
)
}
\arguments{
\item{X}{data containing missing values denoted with NA's.}

\item{imputation_func}{an imputing function. If \code{NULL}, please provide
imputed datasets \code{X_imp} and \code{m}.}

\item{X_imp}{a list of imputed datasets. If \code{NULL} it will be obtained
using \code{imputation_func}.}

\item{m}{the number of multiple imputations to consider, default to 5.}

\item{n_proj}{an integer specifying the number of projections to consider
for the score.}

\item{n_trees_per_proj}{an integer, the number of trees per projection.}

\item{min_node_size}{the minimum number of nodes in a tree.}

\item{n_cores}{an integer, the number of cores to use.}

\item{projection_function}{a function providing the user-specific projections.}

\item{...}{used for compatibility}
}
\value{
numeric value of the score obtained for provided imputation method.
}
\description{
Compute the imputation KL-based scoring rules
}
\examples{
set.seed(111)
X <- random_mcar_data(100, 3, 0.2)
imputation_func <- exp_imputation
DR_IScore(X, imputation_func, m = 2, n_proj = 10, n_trees_per_proj = 2 )


}
\references{
This method is described in detail in:

Näf, Jeffrey, Meta-Lina Spohn, Loris Michel, and Nicolai Meinshausen. 2022.
“Imputation Scores.” https://arxiv.org/abs/2106.03742.
}
