% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_dist.R
\name{edistance}
\alias{edistance}
\title{Energy distance}
\usage{
edistance(X, X_imp, rescale = FALSE)
}
\arguments{
\item{X}{a complete original dataset (without missing values).}

\item{X_imp}{an imputed dataset}

\item{rescale}{a logical, indicating whether the returned value should be
rescaled. Default to \code{FALSE}. See "details" section for more information.}
}
\value{
A numeric value giving the energy distance between the original
dataset and the imputed dataset.
}
\description{
Calculating energy distance/statistic.
}
\details{
This function uses the \link[energy]{eqdist.e} function. According
to this implementation, by default, the function returns the energy statistic
which is given by
\deqn{E(X, Y) = \frac{nm}{n + m} \hat{\varepsilon}{(X, Y)},}
where \eqn{\hat{\varepsilon}{(X, Y)}} is the raw energy distance. To
obtain raw energy distance use \code{rescale = TRUE}.
}
\examples{
X <- matrix(rnorm(100), nrow = 25)
X_imp <- matrix(rnorm(100), nrow = 25)
edistance(X, X_imp)

}
