% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-Score.R
\name{energy_IScore}
\alias{energy_IScore}
\title{Calculates Imputation Score for imputation function}
\usage{
energy_IScore(
  X,
  imputation_func,
  X_imp = NULL,
  multiple = TRUE,
  N = 50,
  max_length = NULL,
  skip_if_needed = TRUE,
  scale = FALSE,
  n_cores = 1,
  silent = TRUE
)
}
\arguments{
\item{X}{data containing missing values denoted with NA's.}

\item{imputation_func}{a function that imputes data.}

\item{X_imp}{imputed dataset of the same size as \code{X}. It's \code{NULL}
by default meaning that it will be obtained by imputation of  \code{X} using
the \code{imputation_func}.}

\item{multiple}{a logical indicating whether provided imputation method is a
multiple imputation approach (i.e. it generates different values to impute
for each call). Default to TRUE. Note that if multiple equals to FALSE, N is
automatically set to 1.}

\item{N}{a numeric value. Number of samples from imputation distribution H.
Default to 50.}

\item{max_length}{Maximum number of variables \eqn{X_j} to consider, can
speed up the code. Default to \code{NULL} meaning that all the columns will
be taken under consideration.}

\item{skip_if_needed}{logical, indicating whether some observations should be
skipped to obtain complete columns for scoring. If FALSE, NA will be returned
for column with no observed variable for training.}

\item{scale}{a logical value. If TRUE, each variable is scaled in the score.}

\item{n_cores}{a number of cores for parallelization.}

\item{silent}{logical indicating whether warnings and messages should be
printed.}
}
\value{
a numerical value denoting weighted Imputation Score obtained for
provided imputation function and a table with scores and weights calculated
for particular columns.
}
\description{
Calculates Imputation Score for imputation function
}
\details{
This function relies on functions \link[Iscores]{energy_Iscore_num}
and \link[Iscores]{energy_Iscore_cat}. Depending on the presence of factor-type
data, these functions compute a score either for purely numerical data or for
mixed data types.

If you want to compute the score for numerical data, make sure that the
dataset does not contain any factor-type variables.

If you want to compute the score for categorical data, ensure that all
categorical variables are preserved as factors.

If your imputation method does not support categorical variables represented
as factors, implement a wrapper function that handles the appropriate data
type conversions before and after imputation.
}
\examples{
set.seed(111)
X <- random_mcar_data(100, 4)
imputation_func <- exp_imputation
energy_IScore(X, imputation_func)

X <-  random_mcar_mixed_data(100, 4, 2)
imputation_func <- median_mode_imputation
energy_IScore(X, imputation_func)

}
\references{
Näf, J., Grzesiak, K., and Scornet, E. (2025).
How to rank imputation methods?
arXiv preprint.
\doi{10.48550/arXiv.2507.11297}.
}
