% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-I-Score_cat.R
\name{energy_Iscore_cat}
\alias{energy_Iscore_cat}
\title{energy-I-Score for imputation of categorical (mixed) data}
\usage{
energy_Iscore_cat(
  X,
  imputation_func,
  X_imp = NULL,
  multiple = TRUE,
  N = 50,
  max_length = NULL,
  skip_if_needed = TRUE
)
}
\arguments{
\item{X}{data containing missing values denoted with NA's.}

\item{imputation_func}{a function that imputes data.}

\item{X_imp}{imputed dataset.}

\item{multiple}{a logical indicating whether provided imputation method is a
multiple imputation approach (i.e. it generates different values to impute
for each call). Default to TRUE. Note that if multiple equals to FALSE, N is
automatically set to 1.}

\item{N}{a numeric value. Number of samples from imputation distribution H.
Default to 50.}

\item{max_length}{Maximum number of variables \eqn{X_j} to consider, can
speed up the code. Default to \code{NULL} meaning that all the columns will
be taken under consideration.}

\item{skip_if_needed}{logical, indicating whether some observations should be
skipped to obtain complete columns for scoring. If FALSE, NA will be returned
for column with no observed variable for training.}
}
\value{
a numerical value denoting weighted Imputation Score obtained for
provided imputation function and a table with scores and weights calculated
for particular columns.
}
\description{
energy-I-Score for imputation of categorical (mixed) data
}
\details{
The categorical variables should be stored as factors. If you need additional
conversion of the data (for example one-hot encoding) for imputation, please,
implement everything within \code{imputation_func} parameter. You can use
\code{miceDRF:::onehot_to_factor} and \code{miceDRF:::factor_to_onehot}
functions.
}
\examples{
set.seed(123)

}
\references{
This method is described in detail in:

Näf, J., Grzesiak, K., and Scornet, E. (2025a). How to rank imputation
methods? arXiv preprint arXiv:2507.11297
(\url{https://doi.org/10.48550/arXiv.2507.11297}).
}
