% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-I-Score_cat.R
\name{factor_to_numeric}
\alias{factor_to_numeric}
\title{Internal function for changing factors to numerical}
\usage{
factor_to_numeric(factor_col)
}
\arguments{
\item{factor_col}{a factor column}
}
\description{
A supplementary function for data management
}
\details{
This function converts factor variables to numeric variables.
}
\keyword{internal}
