% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy-I-Score_cat.R
\name{factor_to_onehot}
\alias{factor_to_onehot}
\title{One hot encoding}
\usage{
factor_to_onehot(dat)
}
\arguments{
\item{dat}{a data containing some factor but numeric columns.}
}
\description{
A supplementary function for one-hot encoding
}
\details{
This function converts factor variables into one-hot encoding
}
\keyword{internal}
