% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_na_data.R
\name{median_mode_imputation}
\alias{median_mode_imputation}
\title{Median/mode imputation}
\usage{
median_mode_imputation(X_miss)
}
\arguments{
\item{X_miss}{A data set containing missing values.}
}
\value{
A completed data set with all missing values imputed.
}
\description{
Imputes numerical variables using their median and categorical variables
using their most frequent observed category.
}
\examples{
X <- random_mcar_mixed_data(100, 3, n_fac = 1)
X_imp <- median_mode_imputation(X)

}
