% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR-I-Score.R
\name{merge_singleton_patterns}
\alias{merge_singleton_patterns}
\title{Merge singleton missingness patterns}
\usage{
merge_singleton_patterns(patterns, groups, ind_singletons)
}
\arguments{
\item{patterns}{A numeric matrix where each row represents a unique
missingness pattern.}

\item{groups}{A list of integer vectors. Each element contains the indices of
observations corresponding to a given pattern in \code{patterns}.}

\item{ind_singletons}{An integer vector indicating indices of patterns in
\code{patterns} that occur only once.}
}
\value{
A list with two elements:
\describe{
  \item{patterns}{Updated matrix of unique missingness patterns.}
  \item{groups}{Updated list of observation indices grouped by pattern.}
}
}
\description{
Merges missingness patterns that occur only once (singleton patterns) into a
single pattern. If the merged pattern already exists among the current
patterns, the corresponding groups of observations are combined. Otherwise,
a new pattern is created and appended.
}
\keyword{internal}
