% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{predict.crf}
\alias{predict.crf}
\title{Prediction for crf}
\usage{
\method{predict}{crf}(
  object,
  Z,
  cutoff = NULL,
  return.regions = FALSE,
  alpha = 0.05,
  combined = FALSE
)
}
\arguments{
\item{object}{a crf object}

\item{Z}{a matrix of candidate points}

\item{cutoff}{the cutoff for the ensemble}

\item{return.regions}{should the individual indicators be returned?}

\item{alpha}{the level}

\item{combined}{the option to have the forests combined or not}
}
\value{
a list of the predictions from a crf object.
}
\description{
Prediction for crf
}
