% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_na_data.R
\name{random_mcar_data}
\alias{random_mcar_data}
\title{Generate random data with MCAR missing values}
\usage{
random_mcar_data(n, p, ratio = 0.2)
}
\arguments{
\item{n}{Number of observations.}

\item{p}{Number of numerical variables.}

\item{ratio}{Proportion of entries to replace with missing values.}
}
\value{
A data frame with \code{n} rows and \code{p} numerical variables
containing missing values.
}
\description{
Generates a numerical dataset consisting of independent standard normal
variables and introduces missing values according to a Missing Completely
at Random (MCAR) mechanism.
}
\examples{
X <- random_mcar_data(100, 3, ratio = 0.2)
head(X)

}
