% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_na_data.R
\name{random_mcar_mixed_data}
\alias{random_mcar_mixed_data}
\title{Generate random mixed data with MCAR missing values}
\usage{
random_mcar_mixed_data(n, p, n_fac = 1, ratio = 0.2)
}
\arguments{
\item{n}{Number of observations.}

\item{p}{Number of numerical variables.}

\item{n_fac}{Number of categorical variables.}

\item{ratio}{Proportion of entries to replace with missing values.}
}
\value{
A data frame containing \code{p} numerical variables and
\code{n_fac} factor variables with missing values.
}
\description{
Generates a mixed dataset containing independent standard normal variables
and categorical variables, then introduces missing values according to a
Missing Completely at Random (MCAR) mechanism.
}
\examples{
X <- random_mcar_mixed_data(100, 3, n_fac = 2, ratio = 0.2)
str(X)

}
