% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sample.vars.proj}
\alias{sample.vars.proj}
\title{Sampling of Projections}
\usage{
sample.vars.proj(ids.x.na, X, projection.function = NULL, normal.proj = T)
}
\arguments{
\item{ids.x.na}{a vector of indices corresponding to NA in the given missingness pattern.}

\item{X}{a matrix of the observed data containing missing values.}

\item{projection.function}{a function providing the user-specific projections.}

\item{normal.proj, }{a boolean, if TRUE, sample from the NA of the pattern and
additionally from the non NA. If FALSE, sample only from the NA of the pattern.}
}
\value{
a vector of variables corresponding to the projection.
}
\description{
Sampling of Projections
}
