% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR-I-Score.R
\name{sample_vars_proj}
\alias{sample_vars_proj}
\title{Sampling of Projections}
\usage{
sample_vars_proj(ids_x_na, X, projection_function = NULL, normal_proj = TRUE)
}
\arguments{
\item{ids_x_na}{a vector of indices corresponding to NA in the given
missingness pattern.}

\item{X}{a matrix of the observed data containing missing values.}

\item{projection_function}{a function providing the user-specific projections.}

\item{normal_proj}{a boolean, if TRUE, sample from the NA of the pattern and
additionally from the non-NA. If FALSE, sample only from the NA of the
pattern.}
}
\value{
a vector of variables corresponding to the projection.
}
\description{
Sampling of Projections
}
\keyword{internal}
