% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{truncProb}
\alias{truncProb}
\title{Truncation of probability}
\usage{
truncProb(p)
}
\arguments{
\item{p}{a numeric value between 0 and 1 to be truncated}
}
\value{
a numeric value, the truncated probability.
}
\description{
Truncation of probability
}
