% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{class.balancing}
\alias{class.balancing}
\title{Balancing of Classes}
\usage{
class.balancing(X.proj.complete, Y.proj, drawA, Xhat, ids.with.missing, vars)
}
\arguments{
\item{X.proj.complete}{matrix with complete projected observations.}

\item{Y.proj}{matrix with projected imputed observations.}

\item{drawA}{vector of indices corresponding to current missingness pattern.}

\item{Xhat}{matrix of full imputed observations.}

\item{ids.with.missing}{vector of indices of observations with missing values.}

\item{vars}{vectors of variables in projection.}
}
\value{
a list of new X.proj.complete and Y.proj.
}
\description{
Balancing of Classes
}
