% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iscores.R, R/helpers.R
\name{combine2Forests}
\alias{combine2Forests}
\title{Combined projection forests}
\usage{
combine2Forests(mod1, mod2)

combine2Forests(mod1, mod2)

combine2Forests(mod1, mod2)
}
\arguments{
\item{mod1}{first forest}

\item{mod2}{second forest}
}
\description{
Combined projection forests

Combined projection forests

Combining projection forests
}
