% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{combine2Forests}
\alias{combine2Forests}
\title{Combining projection forests}
\usage{
combine2Forests(mod1, mod2)
}
\arguments{
\item{mod1}{first forest}

\item{mod2}{second forest}
}
\value{
a new forest combining the first and the second forest
}
\description{
Combining projection forests
}
