% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{combineForests}
\alias{combineForests}
\title{Combining a list of forest}
\usage{
combineForests(list.rf)
}
\arguments{
\item{list.rf}{a list of forests}
}
\value{
a forest combination of the forests stored in list.rf
}
\description{
Combining a list of forest
}
