% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-Score.R
\name{compare_Iscores}
\alias{compare_Iscores}
\title{Calculates IScores for multiple imputation functions}
\usage{
compare_Iscores(X, imputation_list, ...)
}
\arguments{
\item{X}{data containing missing values denoted with NA's.}

\item{imputation_list}{a named list of imputing functions.}
}
\value{
a vector of IScores for provided methods
}
\description{
Calculates IScores for multiple imputation functions
}
\examples{
set.seed(111)
X <- Iscores:::random_mcar_data(100, 4, 0.4)
imputation_list <- list(exp = Iscores:::exp_imputation,
                       norm = Iscores:::norm_imputation)
compare_Iscores(X, imputation_list)

}
