% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_drScore}
\alias{compute_drScore}
\title{compute the density ratio score}
\usage{
compute_drScore(object, Z = Z, multiclass = FALSE, type = "simple")
}
\arguments{
\item{object}{a crf object}

\item{Z}{a matrix of candidate points}

\item{multiclass}{a boolean}

\item{type}{a character}
}
\value{
a numeric value, the score.s
}
\description{
compute the density ratio score
}
