% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{densityRatioScore}
\alias{densityRatioScore}
\title{Computation of the density ratio score}
\usage{
densityRatioScore(
  X,
  Xhat,
  x = NULL,
  X.true = NULL,
  num.proj = 10,
  num.trees.per.proj = 1,
  projection.function = NULL,
  min.node.size = 1,
  multiclass = FALSE,
  compute.glm = FALSE,
  ...
)
}
\arguments{
\item{X}{a matrix of the observed data containing missing values}

\item{Xhat}{a  matrix of imputations having same size as X}

\item{x}{pattern of missing values}

\item{X.true}{the actual true data if available}

\item{num.proj}{an integer specifying the number of projections to consider for the score.}

\item{num.trees.per.proj}{an integer, the number of trees per projection.}

\item{projection.function}{a function providing the possible projections}

\item{min.node.size}{the minimum number of nodes in a tree.}

\item{multiclass}{a boolean}

\item{compute.glm}{a boolean}

\item{...}{additional parameters#'}
}
\value{
a crf object
}
\description{
Computation of the density ratio score
}
