% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{densityRatioScore}
\alias{densityRatioScore}
\title{Computation of the density ratio score}
\usage{
densityRatioScore(
  X,
  Xhat,
  x = NULL,
  num.proj = 10,
  num.trees.per.proj = 1,
  projection.function = NULL,
  min.node.size = 1,
  normal.proj = T
)
}
\arguments{
\item{X}{a matrix of the observed data containing missing values.}

\item{Xhat}{a  matrix of imputations having same size as X.}

\item{x}{pattern of missing values.}

\item{num.proj}{an integer specifying the number of projections.}

\item{num.trees.per.proj}{an integer, the number of trees per projection.}

\item{projection.function}{a function providing the user-specific projections.}

\item{min.node.size}{the minimum number of observations in a leaf of a tree.}

\item{normal.proj, }{a boolean, if TRUE, sample from the NA of the pattern and
additionally from the non NA. If FALSE, sample only from the NA of the pattern.}
}
\value{
a fitted random forest based on random projections
}
\description{
Computation of the density ratio score
}
