% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-Score.R
\name{energy_IScore}
\alias{energy_IScore}
\title{Calculates Imputation Score for imputation function}
\usage{
energy_IScore(X, imputation_func, X_imp = NULL, ...)
}
\arguments{
\item{X}{data containing missing values denoted with NA's.}

\item{imputation_func}{a function that imputes data.}

\item{X_imp}{imputed dataset of the same size as \code{X}. It's \code{NULL}
by default meaning that it will be obtained by imputation of  \code{X} using
the \code{imputation_func}.}

\item{...}{other arguments for the functions \link[Iscores]{energy_Iscore_num}
and \link[Iscores]{energy_Iscore_cat}.}
}
\value{
a numerical value denoting weighted Imputation Score obtained for
provided imputation function and a table with scores and weights calculated
for particular columns.
}
\description{
Calculates Imputation Score for imputation function
}
\details{
This function relies on functions \link[Iscores]{energy_Iscore_num}
and \link[Iscores]{energy_Iscore_cat}. Depending on the presence of factor-type
data, these functions compute a score either for purely numerical data or for
mixed data types.

If you want to compute the score for numerical data, make sure that the
dataset does not contain any factor-type variables.

If you want to compute the score for categorical data, ensure that all
categorical variables are preserved as factors.

If your imputation method does not support categorical variables represented
as factors, implement a wrapper function that handles the appropriate data
type conversions before and after imputation.
}
\examples{
set.seed(111)
X <- Iscores:::random_mcar_data(100, 4)
imputation_func <- Iscores:::exp_imputation
energy_IScore(X, imputation_func)

X <-  Iscores:::random_mcar_mixed_data(100, 4, 2)
imputation_func <- Iscores:::median_mode_imputation
energy_IScore(X, imputation_func)

}
