% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_na_data.R
\name{median_mode_imputation}
\alias{median_mode_imputation}
\title{Standard normal imputation}
\usage{
median_mode_imputation(X_miss)
}
\arguments{
\item{X_miss}{a dataset with missing values}
}
\description{
This is an example function that imputes from N(0, 1)
}
\keyword{internal}
