% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_na_data.R
\name{norm_imputation}
\alias{norm_imputation}
\title{Standard exponential imputation}
\usage{
norm_imputation(X_miss)
}
\arguments{
\item{X_miss}{a dataset with missing values}
}
\description{
This is an example function that imputes from univariate Exp(1) distribution.
}
\keyword{internal}
