% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_na_data.R
\name{random_mcar_mixed_data}
\alias{random_mcar_mixed_data}
\title{Create random example data for imputation}
\usage{
random_mcar_mixed_data(n, p, n_fac = 1, ratio = 0.2)
}
\arguments{
\item{n}{number of samples}

\item{p}{number of variables}

\item{ratio}{fraction of missing values to generate}
}
\description{
This function generates some independent normal variables with missing values
according to MCAR (Missing Completely at Random) mechanism.
}
\keyword{internal}
